Launcher files reorganized
