+++ Memoria.Launcher/Languages/Lang.cs
@@ -23,7 +23,7 @@ namespace Memoria.Launcher
                 XmlElement def = XmlHelper.LoadEmbadedDocument(assembly, $"Languages.en.xml");
                 XmlElement cur = null;
 
-                IniFile iniFile = new IniFile(GameSettingsControl.IniPath);
+                IniFile iniFile = new IniFile(SettingsGrid_Vanilla.IniPath);
                 String forcedLang = iniFile.ReadValue("Memoria", "LauncherLanguage");
 
                 String[] fileNames = String.IsNullOrEmpty(forcedLang) ?
@@ -218,17 +218,6 @@ namespace Memoria.Launcher
             public static readonly string tetraMasterReduceRandomBox0 = GetSettings(nameof(tetraMasterReduceRandomBox0));
             public static readonly string tetraMasterReduceRandomBox1 = GetSettings(nameof(tetraMasterReduceRandomBox1));
             public static readonly string tetraMasterReduceRandomBox2 = GetSettings(nameof(tetraMasterReduceRandomBox2));
-
-            // Advanced Settings
-            /* Unused for now
-            public static readonly string EnableCustomShader = GetSettings(nameof(EnableCustomShader));
-            public static readonly string Shader_Field_Realism = GetSettings(nameof(Shader_Field_Realism));
-            public static readonly string Shader_Field_Toon = GetSettings(nameof(Shader_Field_Toon));
-            public static readonly string Shader_Field_Outlines = GetSettings(nameof(Shader_Field_Outlines));
-            public static readonly string Shader_Battle_Realism = GetSettings(nameof(Shader_Battle_Realism));
-            public static readonly string Shader_Battle_Toon = GetSettings(nameof(Shader_Battle_Toon));
-            public static readonly string Shader_Battle_Outlines = GetSettings(nameof(Shader_Battle_Outlines));
-            */
 
             // Settings.ini Tooltips
             public static readonly string ActiveMonitor_Tooltip = GetSettings(nameof(ActiveMonitor_Tooltip));
