+++ Memoria.Launcher/Launcher/AdvOptionsWindow.xaml
@@ -41,8 +41,8 @@
             HorizontalAlignment="Center"
             VerticalAlignment="Top" />
 
-        <launcher:ShaderSettingControl
-            x:Name="GameSettings"
+        <launcher:SettingsGrid_Shaders
+            x:Name="SettingsGrid_Shaders"
             Width="260"
             Height="157"
             Margin="614,0,0,68"
