+++ Memoria.Launcher/Launcher/UiLauncherButton.xaml
