+++ Memoria.Launcher/Launcher/UiLauncherButton.xaml.cs
