+++ Memoria.Launcher/Launcher/UiModManagerButton.xaml
