+++ Memoria.Launcher/Launcher/UiModManagerButton.xaml.cs
