+++ Memoria.Launcher/Memoria.Launcher.csproj
@@ -46,7 +46,7 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <PropertyGroup>
-    <ApplicationIcon>ff9.ico</ApplicationIcon>
+    <ApplicationIcon>Images\ff9.ico</ApplicationIcon>
   </PropertyGroup>
   <PropertyGroup>
     <StartupObject>Memoria.Launcher.Program</StartupObject>
@@ -100,47 +100,41 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AdvOptionsWindow.cs" />
-    <Compile Include="Extractor.cs" />
+    <Compile Include="Launcher\AdvOptionsWindow.cs" />
     <Compile Include="Languages\Lang.cs" />
     <Compile Include="Languages\XmlDocumentExm.cs" />
     <Compile Include="Languages\XmlHelper.cs" />
-    <Compile Include="Memoria\MemoriaIniCheatControl.cs" />
-    <Compile Include="Memoria\MemoriaIniControl.cs" />
-    <Compile Include="Memoria\Mod.cs" />
-    <Compile Include="Memoria\ShaderSettingControl.cs" />
-    <Compile Include="Memoria\ThrottledWebClient.cs">
+    <Compile Include="Launcher\SettingsGrid_Cheats.cs" />
+    <Compile Include="Launcher\SettingsGrid_Main.cs" />
+    <Compile Include="ModManager\Mod.cs" />
+    <Compile Include="Launcher\SettingsGrid_Shaders.cs" />
+    <Compile Include="ModManager\ThrottledWebClient.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="Memoria\UI.cs" />
-    <Compile Include="Memoria\UiLauncherAdvOptionsCloseButton.cs" />
-    <Compile Include="Memoria\UiLauncherAdvOptionsButton.cs" />
-    <Compile Include="Memoria\UiLauncherModelViewerButton.cs" />
-    <Compile Include="Memoria\UiModManagerButton.xaml.cs">
+    <Compile Include="Launcher\UI.cs" />
+    <Compile Include="Launcher\UiModManagerButton.xaml.cs">
       <DependentUpon>UiModManagerButton.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Memoria\UiLauncherModManagerButton.cs" />
-    <Compile Include="Memoria\ScreenInterrogatory.cs" />
-    <Compile Include="Memoria\FrameworkElementExm.cs" />
     <Compile Include="App.cs" />
     <Compile Include="MainWindow.cs" />
     <Compile Include="Ini\IniFile.cs" />
     <Compile Include="AssemblyInfo.cs" />
-    <Compile Include="Memoria\GameSettingsControl.cs" />
-    <Compile Include="Memoria\JunctionPoint.cs" />
-    <Compile Include="Memoria\UiLauncherPlayButton.cs" />
-    <Compile Include="Memoria\UiLauncherButton.xaml.cs">
+    <Compile Include="Launcher\SettingsGrid_Vanilla.cs" />
+    <Compile Include="Launcher\Utils.cs" />
+    <Compile Include="Launcher\UiLauncherPlayButton.cs" />
+    <Compile Include="Launcher\UiLauncherButton.xaml.cs">
       <DependentUpon>UiLauncherButton.xaml</DependentUpon>
     </Compile>
-    <Compile Include="ModConflictWindow.cs" />
-    <Compile Include="ModManagerWindow.cs" />
+    <Compile Include="ModManager\ModConflictWindow.cs" />
+    <Compile Include="ModManager\ModManagerWindow.cs" />
+    <Compile Include="ModManager\ModManagerUtils.cs" />
     <Compile Include="Program.cs" />
   </ItemGroup>
   <ItemGroup>
     <Resource Include="Images\pc_launcher_bg.png" />
   </ItemGroup>
   <ItemGroup>
-    <Page Include="AdvOptionsWindow.xaml">
+    <Page Include="Launcher\AdvOptionsWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
@@ -148,19 +142,19 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
-    <Page Include="Memoria\UiModManagerButton.xaml">
+    <Page Include="Launcher\UiModManagerButton.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
-    <Page Include="Memoria\UiLauncherButton.xaml">
+    <Page Include="Launcher\UiLauncherButton.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
-    <Page Include="ModConflictWindow.xaml">
+    <Page Include="ModManager\ModConflictWindow.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="ModManagerWindow.xaml">
+    <Page Include="ModManager\ModManagerWindow.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -174,7 +168,6 @@
     <Resource Include="Fonts\centurygothic.ttf" />
     <Resource Include="Fonts\notoemoji.ttf" />
     <Resource Include="Fonts\overpass.ttf" />
-    <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
     <Resource Include="Images\btnReorganize.png" />
@@ -203,7 +196,7 @@
     <Resource Include="Images\btnMoveUpimg.png" />
     <Resource Include="Images\btnUninstallimg.png" />
     <None Include="Catalogs\MemoriaCatalog.xml" />
-    <Content Include="ff9.ico" />
+    <Content Include="Images\ff9.ico" />
     <Resource Include="Images\new_launcher_bg.png" />
   </ItemGroup>
   <ItemGroup>
@@ -223,6 +216,7 @@
       <Version>0.38.0</Version>
     </PackageReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="AfterResolveReferences">
     <ItemGroup>
