--- Memoria.Launcher/Memoria/UiLauncherAdvOptionsButton.cs
@@ -1,24 +1,0 @@
-using System;
-using System.Threading.Tasks;
-using System.Windows;
-
-namespace Memoria.Launcher
-{
-    public sealed class UiLauncherAdvOptionsButton : UiModManagerButton
-    {
-        public UiLauncherAdvOptionsButton()
-        {
-            Label = Lang.Launcher.AdvSettings;
-        }
-
-        protected override async Task DoAction()
-        {
-            MainWindow mainWindow = (MainWindow)this.GetRootElement();
-            if (mainWindow.AdvOptionsWindow == null)
-                mainWindow.AdvOptionsWindow = new AdvOptionsWindow();
-            mainWindow.AdvOptionsWindow.Owner = mainWindow;
-            mainWindow.AdvOptionsWindow.Show();
-            mainWindow.AdvOptionsWindow.Activate();
-        }
-    }
-}
