--- Memoria.Launcher/Memoria/UiLauncherModManagerButton.cs
@@ -1,24 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-using System.Windows;
-
-namespace Memoria.Launcher
-{
-    public sealed class UiLauncherModManagerButton : UiModManagerButton
-    {
-        public UiLauncherModManagerButton()
-        {
-            Label = Lang.Launcher.ModManager;
-        }
-
-        protected override async Task DoAction()
-        {
-            MainWindow mainWindow = (MainWindow)this.GetRootElement();
-            if (mainWindow.ModdingWindow == null)
-                mainWindow.ModdingWindow = new ModManagerWindow();
-            mainWindow.ModdingWindow.Owner = mainWindow;
-            mainWindow.ModdingWindow.Show();
-            mainWindow.ModdingWindow.Activate();
-        }
-    }
-}
