+++ Memoria.Launcher/ModManager/Mod.cs
