+++ Memoria.Launcher/ModManager/ModConflictWindow.xaml
