+++ Memoria.Launcher/ModManager/ModManagerWindow.xaml
@@ -5,7 +5,7 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:launcher="clr-namespace:Memoria.Launcher" 
-    d:DataContext="{d:DesignInstance Type=launcher:MemoriaIniControl}"
+    d:DataContext="{d:DesignInstance Type=launcher:SettingsGrid_Main}"
     Title="{Binding WindowTitle, FallbackValue='Memoria Mod Manager'}"
     Width="1100"
     Height="720"
