+++ Memoria.Launcher/ModManager/ThrottledWebClient.cs
